Time: 43 ms (13.29%), Space: 88.1 MB (75.7%) - LeetHub

+++ 1335-maximum-candies-allocated-to-k-children/1335-maximum-candies-allocated-to-k-children.cpp
@@ -3,7 +3,7 @@ public:
     bool ok(int mid, vector<int>& candies, long long k)
     {
         long long cnt=0;
-        long long  sum=0;
+        long   sum=0;
         for (int& val : candies)
             {cnt += val/mid;
             sum+=val;
